backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
networkx
python-dotenv
